IGNITE-16566 Move distribution zones exception to core module (temporary solution)
