+++ modules/core/src/main/java/org/apache/ignite/internal/distributionzones/DistributionZoneAlreadyExistsException.java
@@ -15,12 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.lang;
+package org.apache.ignite.internal.distributionzones;
 
 import static org.apache.ignite.lang.ErrorGroups.DistributionZones.ZONE_ALREADY_EXISTS_ERR;
 
 import java.util.UUID;
+import org.apache.ignite.lang.IgniteInternalException;
 
+// TODO
 /**
  * This exception is thrown when a new distribution zone failed to be created,
  * because a distribution zone with same name already exists.
