+++ modules/core/src/main/java/org/apache/ignite/internal/distributionzones/DistributionZoneBindTableException.java
@@ -15,43 +15,39 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.lang;
+package org.apache.ignite.internal.distributionzones;
 
-import static org.apache.ignite.lang.ErrorGroups.DistributionZones.ZONE_NOT_FOUND_ERR;
+import static org.apache.ignite.lang.ErrorGroups.DistributionZones.ZONE_DROP_ERR;
 
 import java.util.UUID;
+import org.apache.ignite.lang.IgniteInternalException;
 import org.jetbrains.annotations.Nullable;
 
+// TODO
 /**
- * Exception is thrown when appropriate distribution zone can`t be found.
+ * Exception is thrown when the distribution zone cannot be dropped because there is a table bound to the distribution zone.
  */
-public class DistributionZoneNotFoundException extends IgniteInternalException {
+public class DistributionZoneBindTableException extends IgniteInternalException {
     /**
      * The constructor.
      *
      * @param zoneName Zone name.
+     * @param tableName Table name.
      */
-    public DistributionZoneNotFoundException(String zoneName) {
-        this(zoneName, null);
-    }
-
-    /**
-     * The constructor.
-     *
-     * @param zoneId Zone id.
-     */
-    public DistributionZoneNotFoundException(int zoneId) {
-        super(ZONE_NOT_FOUND_ERR, "Distribution zone is not found [zoneId=" + zoneId + ']');
+    public DistributionZoneBindTableException(String zoneName, String tableName) {
+        this(zoneName, tableName, null);
     }
 
     /**
      * The constructor.
      *
      * @param zoneName Zone name.
+     * @param tableName Table name.
      * @param cause Optional nested exception (can be {@code null}).
      */
-    public DistributionZoneNotFoundException(String zoneName, @Nullable Throwable cause) {
-        super(ZONE_NOT_FOUND_ERR, "Distribution zone is not found [zoneName=" + zoneName + ']', cause);
+    public DistributionZoneBindTableException(String zoneName, String tableName, @Nullable Throwable cause) {
+        super(ZONE_DROP_ERR, "Distribution zone is assigned to the table [zoneName=" + zoneName + ", tableName=" + tableName + ']',
+                cause);
     }
 
     /**
@@ -62,7 +58,7 @@ public class DistributionZoneNotFoundException extends IgniteInternalException {
      * @param message Error message.
      * @param cause Cause exception.
      */
-    public DistributionZoneNotFoundException(UUID traceId, int code, String message, Throwable cause) {
+    public DistributionZoneBindTableException(UUID traceId, int code, String message, Throwable cause) {
         super(traceId, code, message, cause);
     }
 }
